Add schema for Summary
- Add error check for duplicate values for columns
- Refactor TabularInputs
