+++ src/redux/selectors/domainData/props-with-state.js
@@ -1,17 +1,5 @@
 import { createSelector } from 'reselect';
-
-import {
-    getObjectChildren,
-    compareString,
-    decodeDate,
-} from '#rsu/common';
-import { median, sum, bucket } from '#rsu/stats';
-import {
-    requiredCondition,
-    dateCondition,
-} from '#rscg/Faram';
-// FIXME: do not import Baksa
-import Baksa from '#components/input/Baksa';
+import { compareString } from '#rsu/common';
 
 import { activeUserSelector } from '../auth';
 import {
@@ -45,195 +33,6 @@ import {
 
 const emptyList = [];
 const emptyObject = {};
-
-// helpers
-
-const createFieldSchema = (field, shouldBeOptional) => {
-    const {
-        isRequired: isRequiredFromField,
-        fieldType,
-        title,
-    } = field;
-    const isRequired = isRequiredFromField && !shouldBeOptional;
-    switch (fieldType) {
-        case 'date':
-            return isRequired
-                ? [requiredCondition, dateCondition]
-                : [dateCondition];
-        case 'daterange':
-            return {
-                fields: {
-                    from: [dateCondition],
-                    to: [dateCondition],
-                },
-                validation: ({ from, to } = {}) => {
-                    const errors = [];
-                    if (!from && !to && isRequired) {
-                        // FIXME: use strings
-                        errors.push(`Either ${title} Start Date or ${title} End Date is required.`);
-                    }
-                    if (from && to && decodeDate(from) > decodeDate(to)) {
-                        // FIXME: use strings
-                        errors.push(`Invalid ${title} Range`);
-                    }
-                    return errors;
-                },
-            };
-        default:
-            return isRequired
-                ? [requiredCondition]
-                : [];
-    }
-};
-
-const createScoreSchema = (scorePillars = [], scoreMatrixPillars = []) => {
-    const scoreSchema = {
-        fields: {
-            pillars: [],
-            matrixPillars: [],
-
-            finalScore: [],
-        },
-    };
-
-    scorePillars.forEach((pillar) => {
-        scoreSchema.fields[`${pillar.id}-score`] = [];
-    });
-    scoreMatrixPillars.forEach((pillar) => {
-        scoreSchema.fields[`${pillar.id}-matrix-score`] = [];
-    });
-
-    return scoreSchema;
-};
-
-const createAdditionalDocumentsSchema = () => {
-    const {
-        bothPageRequiredCondition,
-        validPageRangeCondition,
-        validPageNumbersCondition,
-        pendingCondition,
-    } = Baksa;
-
-    const schema = { fields: {
-        executiveSummary: [
-            bothPageRequiredCondition,
-            validPageRangeCondition,
-            validPageNumbersCondition,
-            pendingCondition,
-        ],
-        assessmentData: [pendingCondition],
-        questionnaire: [
-            bothPageRequiredCondition,
-            validPageRangeCondition,
-            validPageNumbersCondition,
-            pendingCondition,
-        ],
-    } };
-    return schema;
-};
-
-const createBasicInformationSchema = (aryTemplateMetadata = {}) => {
-    // Dynamic fields from metadataGroup
-    const dynamicFields = {};
-    Object.keys(aryTemplateMetadata).forEach((key) => {
-        aryTemplateMetadata[key].fields.forEach((field) => {
-            dynamicFields[field.id] = createFieldSchema(field);
-        });
-    });
-
-    const schema = { fields: dynamicFields };
-    return schema;
-};
-
-// NOTE:
-export const getDataCollectionTechnique = (aryTemplateMethodology) => {
-    let dataCollectionTechnique;
-    aryTemplateMethodology.some(
-        group => group.fields.some((field) => {
-            if (field.title.toLowerCase().trim() === 'data collection technique') {
-                dataCollectionTechnique = field;
-                return true;
-            }
-            return false;
-        }),
-    );
-    return dataCollectionTechnique;
-};
-
-// NOTE:
-export const isSecondaryDataReviewOption = option => (
-    option && option.label.toLowerCase().trim() === 'secondary data review'
-);
-
-const createMethodologySchema = (aryTemplateMethodology = {}) => {
-    const dataCollectionTechnique = getDataCollectionTechnique(aryTemplateMethodology);
-
-    const schema = { fields: {
-        attributes: {
-            keySelector: d => d.key,
-            identifier: (value = {}) => {
-                if (!dataCollectionTechnique) {
-                    return 'default';
-                }
-                const key = value[dataCollectionTechnique.id];
-                if (!key) {
-                    return 'default';
-                }
-                const selectedOption = dataCollectionTechnique.options.find(
-                    option => option.key === key,
-                );
-                return isSecondaryDataReviewOption(selectedOption) ? 'secondaryDataReview' : 'default';
-            },
-            member: {
-                secondaryDataReview: {
-                    fields: {/* NOTE: injected here */},
-                },
-                default: {
-                    fields: {/* NOTE: injected here */},
-                },
-            },
-            validation: (value) => {
-                const errors = [];
-                if (!value || value.length < 1) {
-                    // FIXME: Use strings
-                    errors.push('There should be at least one Collection Technique');
-                }
-                return errors;
-            },
-        },
-
-        sectors: [],
-        focuses: [],
-        locations: [],
-        affectedGroups: [],
-
-        objectives: [],
-        dataCollectionTechniques: [],
-        sampling: [],
-        limitations: [],
-    } };
-
-    const dynamicFields = {};
-    Object.keys(aryTemplateMethodology).forEach((key) => {
-        const methodologyGroup = aryTemplateMethodology[key];
-        methodologyGroup.fields.forEach((field) => {
-            dynamicFields[field.id] = createFieldSchema(field);
-        });
-    });
-    schema.fields.attributes.member.default.fields = dynamicFields;
-
-    const anotherDynamicFields = {};
-    Object.keys(aryTemplateMethodology).forEach((key) => {
-        const methodologyGroup = aryTemplateMethodology[key];
-        methodologyGroup.fields.forEach((field) => {
-            const shouldBeOptional = dataCollectionTechnique.id !== anotherDynamicFields.id;
-            anotherDynamicFields[field.id] = createFieldSchema(field, shouldBeOptional);
-        });
-    });
-    schema.fields.attributes.member.secondaryDataReview.fields = anotherDynamicFields;
-
-    return schema;
-};
 
 
 // active user
@@ -538,91 +337,6 @@ export const assessmentMinFinalScoreSelector = createSelector(
 export const assessmentMaxFinalScoreSelector = createSelector(
     assessmentScoreBucketsSelector,
     buckets => Math.max(...buckets.map(s => s[2])),
-);
-
-export const assessmentSchemaSelector = createSelector(
-    aryTemplateMetadataSelector,
-    aryTemplateMethodologySelector,
-    assessmentPillarsSelector,
-    assessmentMatrixPillarsSelector,
-    (
-        aryTemplateMetadata,
-        aryTemplateMethodology,
-        scorePillars,
-        scoreMatrixPillars,
-    ) => {
-        const schema = { fields: {
-            metadata: {
-                fields: {
-                    basicInformation: createBasicInformationSchema(aryTemplateMetadata),
-                    additionalDocuments: createAdditionalDocumentsSchema(),
-                },
-            },
-            methodology: createMethodologySchema(aryTemplateMethodology),
-            summary: [],
-            score: createScoreSchema(scorePillars, scoreMatrixPillars),
-        } };
-        return schema;
-    },
-);
-
-export const assessmentComputeSchemaSelector = createSelector(
-    assessmentPillarsSelector,
-    assessmentMatrixPillarsSelector,
-    assessmentScoreScalesSelector,
-    assessmentScoreBucketsSelector,
-    (
-        scorePillars = [],
-        scoreMatrixPillars = [],
-        scoreScales = [],
-        scoreBuckets = [],
-    ) => {
-        if (scoreScales.length === 0) {
-            return {};
-        }
-
-        const scoreSchema = {};
-
-        const getScaleVal = v => scoreScales.find(s => String(s.value) === String(v)).value;
-
-        scorePillars.forEach((pillar) => {
-            scoreSchema[`${pillar.id}-score`] = (data, score) => {
-                const pillarObj = getObjectChildren(score, ['pillars', pillar.id]) || emptyObject;
-                const pillarValues = Object.values(pillarObj).map(v => getScaleVal(v));
-                return sum(pillarValues);
-            };
-        });
-
-        scoreMatrixPillars.forEach((pillar) => {
-            const scales = Object.values(pillar.scales).reduce(
-                (acc, b) => [...acc, ...Object.values(b)],
-                [],
-            );
-            const getMatrixScaleVal = v => scales.find(s => String(s.id) === String(v)).value;
-
-            scoreSchema[`${pillar.id}-matrix-score`] = (data, score) => {
-                const pillarObj = getObjectChildren(score, ['matrixPillars', pillar.id]) || emptyObject;
-                const pillarValues = Object.values(pillarObj).map(v => getMatrixScaleVal(v));
-                return median(pillarValues) * 5;
-            };
-        });
-
-        scoreSchema.finalScore = (data, score) => {
-            const pillarScores = scorePillars.map(
-                p => (getObjectChildren(score, [`${p.id}-score`]) || 0) * p.weight,
-            );
-            const matrixPillarScores = scoreMatrixPillars.map(
-                p => (getObjectChildren(score, [`${p.id}-matrix-score`]) || 0) * p.weight,
-            );
-
-            const average = sum([...pillarScores, ...matrixPillarScores]);
-            return bucket(average, scoreBuckets);
-        };
-
-        return { fields: {
-            score: { fields: scoreSchema },
-        } };
-    },
 );
 
 // afIdFromRoute
